Modify in accordance to PR comments
