+++ .github/workflows/bencher.yml
@@ -1,6 +1,5 @@
 on:
   push:
-    branches: john-benchmark
 
 jobs:
   benchmark_base_branch:
